+++ packages/_private/angular-demo/src/demoform/demoform.component.ts
@@ -1,5 +1,5 @@
 import type { SynChangeEvent, SynRange } from '@synergy-design-system/components';
-import { serialize } from '@synergy-design-system/components';
+import { highlightOptionRenderer, serialize } from '@synergy-design-system/components';
 import { Component, ElementRef, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 
@@ -13,6 +13,7 @@ const initialData = {
   gender: '',
   happiness: '5',
   name: '',
+  nationality: '',
   newsletterAngular: false,
   newsletterBeta: false,
   newsletterReact: false,
@@ -37,6 +38,10 @@ export class DemoForm {
   @ViewChild('donationRef') donationRef!: ElementRef<SynRange>;
 
   formData!: FormGroup;
+
+  nationalities: string[] = ['American', 'Australian', 'Brazilian', 'British', 'Canadian', 'Chinese', 'Dutch', 'French', 'German', 'Greek', 'Indian', 'Italian', 'Japanese', 'Korean', 'Mexican', 'Russian', 'Spanish', 'Swedish', 'Turkish'];
+
+  highlightOptionRenderer = highlightOptionRenderer;
 
   private _initFormData() {
     this.formData = this.fb.group({